use a Gradle logger via LoggerAdapter
